Destructure shared gas estimation module
